https://jira.airtel.africa/jira/browse/CST-2072
TC17	Verify that a dropdown is displayed under category option along with all the already created template categories to select from.
TC18	Verify that a dropdown is displayed to select the roles for which the selected templates  will be visible.
TC19	Verify that a dropdown is displayed to select the Agent Channel for which the selected templates  will be visible.
TC20	Verify that clicking on the radio buttion for Add template Category opens up a page wherein new template categories can be created and already created ones can be stored.
TC21	Verify that clicking on the Add template "+" button adds the respective Template category to the list of created templates at the bottom.
TC22	Verify that once the respective template category is created then it must display in the dropdown for template category on the create template Page.